Make HeaderNoImage component reusable and use it on both interior page and Areas to Support page types
